requirements.txt
Pillow
customtkinter
requests

